requirements.txt
# Core Backend
fastapi>=0.104.0
uvicorn[standard]>=0.30
pydantic>=2.8
pydantic-settings>=2.0.3
python-multipart>=0.0.6

# Frontend
streamlit>=1.28.0

# Model Adapters
llama-cpp-python>=0.2.18
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
onnxruntime>=1.16.0
sentencepiece>=0.1.99
protobuf>=4.24.4

# Training & PEFT
peft>=0.6.0
bitsandbytes>=0.41.0
trl>=0.7.4

# Utilities
numpy>=1.26
pandas>=2.0.0
scipy>=1.11
scikit-learn>=1.3.0
safetensors>=0.4.0
httpx>=0.25.0
aiofiles>=23.2.1
pyyaml>=6.0
jsonschema>=4.19.0
tqdm>=4.66.0
networkx>=3.3
duckdb>=1.0.0
faiss-cpu>=1.8.0
orjson>=3.10
cryptography>=43.0
pyjwt>=2.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
requests>=2.31.0

# Diagnostics & Interpretability
einops>=0.7.0

# Optional acceleration (Windows DirectML)
# onnxruntime-directml; sys_platform == 'win32'
